Break out events into own module
